[WIP] amny updates, general abstractions, more to come
