+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 #sdl2 = { version = "0.32.2", features = ["bundled", "static-link"] }
 sdl2 = "0.32.2"
 gl = "0.14.0"
+failure = "0.1"
 
 [build-dependencies]
 walkdir = "2.1"
