+++ src/helpers/shader.rs
@@ -1,0 +1,208 @@
+extern crate gl;
+
+use crate::resources::{self, Resources};
+
+use std::ffi::{CString, CStr};
+
+// http://nercury.github.io/rust/opengl/tutorial/2018/02/10/opengl-in-rust-from-scratch-03-compiling-shaders.html
+
+fn create_whitespace_cstring_with_len(len: usize) -> CString {
+    let mut buffer: Vec<u8> = Vec::with_capacity(len as usize + 1);
+    // fill buffer with len spaces
+    buffer.extend([b' '].iter().cycle().take(len as usize));
+    // convert buffer to CString
+    unsafe { CString::from_vec_unchecked(buffer) }
+}
+
+#[derive(Debug, Fail)]
+pub enum Error {
+    #[fail(display = "Failed to load resource {}", name)]
+    ResourceLoad { name: String, #[cause] inner: resources::Error },
+    #[fail(display = "Can not  determine shader type from resource {}", name)]
+    CanNotDetermineShaderTypeForResource { name: String },
+    #[fail(display = "Failed to compile shader {}: {}", name, message)]
+    CompileError { name: String, message: String },
+    #[fail(display = "Failed to link program {}: {}", name, message)]
+    LinkError { name: String, message: String },
+}
+
+pub struct Program {
+    id: gl::types::GLuint,
+}
+
+impl Program {
+    pub fn from_shaders(shaders: &[Shader]) -> Result<Program, String> {
+        let program_id = unsafe { gl::CreateProgram() };
+        let mut success: gl::types::GLint = 1;
+
+        for shader in shaders {
+            unsafe { gl::AttachShader(program_id, shader.id()); }
+        }
+
+        unsafe {
+            gl::LinkProgram(program_id);
+            gl::GetProgramiv(program_id, gl::LINK_STATUS, &mut success);
+        }
+
+        if success == 0 {
+            let mut len: gl::types::GLint = 0;
+
+            unsafe {
+                gl::GetProgramiv(program_id, gl::INFO_LOG_LENGTH, &mut len);
+            }
+
+            let error = create_whitespace_cstring_with_len(len as usize);
+
+            unsafe {
+                gl::GetProgramInfoLog(
+                    program_id,
+                    len,
+                    std::ptr::null_mut(),
+                    error.as_ptr() as *mut gl::types::GLchar
+                );
+            }
+
+            return Err(error.to_string_lossy().into_owned());
+        }
+
+        for shader in shaders {
+            unsafe {
+                gl::DetachShader(program_id, shader.id());
+            }
+        }
+
+        Ok(Program { id: program_id })
+    }
+
+    pub fn from_resource(res: &Resources, name: &str) -> Result<Program, Error> {
+        const POSSIBLE_EXT: [&str; 2] = [
+            ".vertex",
+            ".fragment",
+        ];
+
+        let shaders = POSSIBLE_EXT.iter()
+            .map(|file_extension| {
+                Shader::from_res(res, &format!("{}{}", name, file_extension))
+            })
+            .collect::<Result<Vec<Shader>, Error>>()?;
+
+        Program::from_shaders(&shaders[..]).map_err(|message| Error::LinkError {
+            name: name.into(),
+            message
+        })
+    }
+
+    pub fn set_used(&self) {
+        unsafe {
+            gl::UseProgram(self.id);
+        }
+    }
+
+    pub fn id(&self) -> gl::types::GLuint {
+        self.id
+    }
+}
+
+pub struct Shader {
+    id: gl::types::GLuint,
+}
+
+impl Drop for Program {
+    fn drop(&mut self) {
+        unsafe {
+            gl::DeleteProgram(self.id);
+        }
+    }
+}
+
+impl Shader {
+    fn from_source(
+        source: &CStr,
+        kind: gl::types::GLenum
+    ) -> Result<Shader, String> {
+        let id = shader_from_source(source, kind)?;
+        Ok(Shader { id })
+    }
+
+    pub fn from_vertex_source(source: &CStr) -> Result<Shader, String> {
+        Shader::from_source(source, gl::VERTEX_SHADER)
+    }
+
+    pub fn from_fragment_source(source: &CStr) -> Result<Shader, String> {
+        Shader::from_source(source, gl::FRAGMENT_SHADER)
+    }
+
+    pub fn from_res(res: &Resources, name: &str) -> Result<Shader, Error> {
+        const POSSIBLE_EXT: [(&str, gl::types::GLenum); 2] = [
+            (".vertex", gl::VERTEX_SHADER),
+            (".fragment", gl::FRAGMENT_SHADER),
+        ];
+        let shader_kind = POSSIBLE_EXT.iter()
+            .find(|&&(file_extension, _)| {
+                name.ends_with(file_extension)
+            })
+            .map(|&(_, kind)| kind)
+            .ok_or_else(|| Error::CanNotDetermineShaderTypeForResource { name: name.into() })?;
+        let source = res.load_cstring(name)
+            .map_err(|e| Error::ResourceLoad {
+                name: name.into(),
+                inner: e,
+            })?;
+
+        Shader::from_source(&source, shader_kind).map_err(|message| Error::CompileError {
+            name: name.into(),
+            message,
+        })
+    }
+
+    pub fn id(&self) -> gl::types::GLuint {
+        self.id
+    }
+}
+
+impl Drop for Shader {
+    fn drop(&mut self) {
+        unsafe {
+            gl::DeleteShader(self.id);
+        }
+    }
+}
+
+pub fn shader_from_source(
+    source: &CStr,
+    kind: gl::types::GLuint
+) -> Result<gl::types::GLuint, String> {
+    let id = unsafe { gl::CreateShader(kind) };
+    let mut success: gl::types::GLint = 1;
+
+    unsafe {
+        gl::ShaderSource(id, 1, &source.as_ptr(), std::ptr::null());
+        gl::CompileShader(id);
+
+        gl::GetShaderiv(id, gl::COMPILE_STATUS, &mut success);
+    }
+
+    if success == 0 {
+        // cont
+        let mut len: gl::types::GLint = 0;
+
+        unsafe {
+            gl::GetShaderiv(id, gl::INFO_LOG_LENGTH, &mut len);
+        }
+
+        let error = create_whitespace_cstring_with_len(len as usize);
+
+        unsafe {
+            gl::GetShaderInfoLog(
+                id,
+                len,
+                std::ptr::null_mut(),
+                error.as_ptr() as *mut gl::types::GLchar
+            );
+        }
+
+        return Err(error.to_string_lossy().into_owned());
+    }
+
+    Ok(id)
+}
