+++ src/main.rs
@@ -1,3 +1,5 @@
+#[macro_use] extern crate failure;
+
 extern crate sdl2;
 extern crate gl;
 
@@ -7,14 +9,47 @@ use std::path::Path;
 pub mod helpers;
 pub mod resources;
 
+use helpers::data;
 use resources::Resources;
 
 const WIDTH: u32 = 720;
 const HEIGHT: u32 = 480;
 
-// http://nercury.github.io/rust/opengl/tutorial/2018/02/12/opengl-in-rust-from-scratch-06-gl-generator.html
+// http://nercury.github.io/rust/opengl/tutorial/2018/02/15/opengl-in-rust-from-scratch-08-failure.html
 
 fn main() {
+    if let Err(e) = run() {
+        println!("{}", failure_to_string(e));
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+#[repr(C, packed)]
+struct Vertex {
+    pos: data::f32_f32_f32,
+    clr: data::f32_f32_f32,
+}
+
+impl Vertex {
+    fn vertex_attrib_pointers() {
+        let stride = std::mem::size_of::<Self>();
+        let location = 0;
+        let offset = 0;
+
+        unsafe {
+            data::f32_f32_f32::vertex_attrib_pointer(stride, location, offset);
+        }
+
+        let location = 1;
+        let offset = offset + std::mem::size_of::<data::f32_f32_f32>();
+
+        unsafe {
+            data::f32_f32_f32::vertex_attrib_pointer(stride, location, offset);
+        }
+    }
+}
+
+fn run() -> Result<(), failure::Error> {
     let sdl = sdl2::init().unwrap();
     let video_subsystem = sdl.video().unwrap();
     let gl_attr = video_subsystem.gl_attr();
@@ -55,11 +90,11 @@ fn main() {
 
     shader_program.set_used();
 
-    let vertices: Vec<f32> = vec![
+    let vertices: Vec<Vertex> = vec![
         // positions        // colors
-        -0.5, -0.5, 0.0,    1.0, 0.0, 0.0,
-        0.5, -0.5, 0.0,     0.0, 1.0, 0.0,
-        0.0, 0.5, 0.0,      0.0, 0.0, 1.0
+       Vertex { pos: (-0.5, -0.5, 0.0).into(), clr: (1.0, 0.0, 0.0).into() }, // bottom right
+       Vertex { pos: (0.5, -0.5, 0.0).into(), clr: (0.0, 1.0, 0.0).into() }, // bottom left
+       Vertex { pos: (0.0, 0.5, 0.0).into(), clr: (0.0, 0.0, 1.0).into() } // top
     ];
 
     let mut vbo: gl::types::GLuint = 0;
@@ -68,7 +103,7 @@ fn main() {
         gl::BindBuffer(gl::ARRAY_BUFFER, vbo);
         gl::BufferData(
             gl::ARRAY_BUFFER, // target
-            (vertices.len() * std::mem::size_of::<f32>()) as gl::types::GLsizeiptr, // size of data in bytes
+            (vertices.len() * std::mem::size_of::<Vertex>()) as gl::types::GLsizeiptr, // size of data in bytes
             vertices.as_ptr() as *const gl::types::GLvoid, // pointer to data
             gl::STATIC_DRAW, //usage
         );
@@ -82,24 +117,9 @@ fn main() {
         gl::GenVertexArrays(1, &mut vao);
         gl::BindVertexArray(vao);
         gl::BindBuffer(gl::ARRAY_BUFFER, vbo);
-        gl::EnableVertexAttribArray(0);
-        gl::VertexAttribPointer(
-            0,
-            3,
-            gl::FLOAT,
-            gl::FALSE,
-            (6 * std::mem::size_of::<f32>()) as gl::types::GLint,
-            std::ptr::null()
-        );
-        gl::EnableVertexAttribArray(1);
-        gl::VertexAttribPointer(
-            1,
-            3,
-            gl::FLOAT,
-            gl::FALSE,
-            (6 * std::mem::size_of::<f32>()) as gl::types::GLint,
-            (3 * std::mem::size_of::<f32>()) as *const gl::types::GLvoid
-        );
+
+        Vertex::vertex_attrib_pointers();
+
         gl::BindBuffer(gl::ARRAY_BUFFER, 0); // unbind buffer
         gl::BindVertexArray(0);
     }
@@ -129,4 +149,31 @@ fn main() {
 
         window.gl_swap_window();
     }
+
+    Ok(())
+}
+
+pub fn failure_to_string(e: failure::Error) -> String {
+    use std::fmt::Write;
+
+    let mut result = String::new();
+
+    for (i, cause) in e.iter_chain().collect::<Vec<_>>().into_iter().rev().enumerate() {
+        if i > 0 {
+            let _ = writeln!(&mut result, " Which caused the following issue:");
+        }
+        let _ = writeln!(&mut result, "{}", cause);
+        if let Some(backtrace) = cause.backtrace() {
+            let backtrace_str = format!("{}", backtrace);
+            if backtrace_str.len() > 0 {
+                let _ = writeln!(&mut result, " This happened at {}", backtrace);
+            } else {
+                let _ = writeln!(&mut result);
+            }
+        } else {
+            let _ = writeln!(&mut result);
+        }
+    }
+
+    result
 }
