+++ src/resources/mod.rs
@@ -3,10 +3,13 @@ use std::fs;
 use std::io::{self, Read};
 use std::ffi;
 
-#[derive(Debug)]
+#[derive(Debug, Fail)]
 pub enum Error {
-    Io(io::Error),
+    #[fail(display = "I/O error")]
+    Io(#[cause] io::Error),
+    #[fail(display = "Failed to read CString from file that contains 0")]
     FileContainsNil,
+    #[fail(display = "Failed to get executable path")]
     FailedToGetExePath,
 }
